nam_sparse_merkle_tree
Module merge

Functions

hash_leaf

hash_leaf = hash(prefix | key | value) zero value indicates a key is to be deleted, this function returns zero for zero value

merge

Merge two hashes this function is optimized for ZERO_HASH if one of lhs or rhs is ZERO_HASH, this function just returns the another one